MNC-1: create projects: Lib and Console
