+++ MatrixNetCore.Lib/AboutVectorFiles.txt
