+++ MatrixNetCore.sln
@@ -3,7 +3,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.6.33723.286
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MatrixNetCore", "MatrixNetCore\MatrixNetCore.csproj", "{E1D6AFD3-A52D-4233-9094-3DF9C6CD08AB}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MatrixNetCore.Console", "MatrixNetCore.Console\MatrixNetCore.Console.csproj", "{CCC83F77-E238-4878-957B-195DDC945093}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MatrixNetCore.Lib", "MatrixNetCore.Lib\MatrixNetCore.Lib.csproj", "{EE0DC4C0-687D-45E8-AC53-D3C4D7C759ED}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -11,10 +13,14 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{E1D6AFD3-A52D-4233-9094-3DF9C6CD08AB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{E1D6AFD3-A52D-4233-9094-3DF9C6CD08AB}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{E1D6AFD3-A52D-4233-9094-3DF9C6CD08AB}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{E1D6AFD3-A52D-4233-9094-3DF9C6CD08AB}.Release|Any CPU.Build.0 = Release|Any CPU
+		{CCC83F77-E238-4878-957B-195DDC945093}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{CCC83F77-E238-4878-957B-195DDC945093}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{CCC83F77-E238-4878-957B-195DDC945093}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{CCC83F77-E238-4878-957B-195DDC945093}.Release|Any CPU.Build.0 = Release|Any CPU
+		{EE0DC4C0-687D-45E8-AC53-D3C4D7C759ED}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{EE0DC4C0-687D-45E8-AC53-D3C4D7C759ED}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{EE0DC4C0-687D-45E8-AC53-D3C4D7C759ED}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{EE0DC4C0-687D-45E8-AC53-D3C4D7C759ED}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
